requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
requests
pandas
